+++ lib/src/providers/implementation/openai_provider.dart
@@ -13,43 +13,35 @@ class OpenAiProvider extends Provider {
   /// Creates a new [OpenAiProvider] with the given parameters.
   ///
   /// The [modelName] is the name of the OpenAI model to use.
-  /// If not provided, [defaultModelName] is used.
+  /// If not provided, [OpenAiModel.defaultModelName] is used.
   /// The [embeddingModelName] is the name of the OpenAI embedding model to use.
-  /// If not provided, [defaultEmbeddingModelName] is used.
+  /// If not provided, [OpenAiModel.defaultEmbeddingModelName] is used.
   /// The [apiKey] is the API key to use for authentication.
   /// If not provided, it's retrieved from the environment.
   OpenAiProvider({
     this.alias,
-    String? modelName,
-    String? embeddingModelName,
+    this.modelName,
+    this.embeddingModelName,
     String? apiKey,
     this.baseUrl,
     this.temperature,
     this.caps = ProviderCaps.all,
-  }) : modelName = modelName ?? defaultModelName,
-       embeddingModelName = embeddingModelName ?? defaultEmbeddingModelName,
-       apiKey = apiKey ?? platform.getEnv(apiKeyName);
-
-  /// The default model name to use if none is provided.
-  static const defaultModelName = 'gpt-4o';
-
-  /// The default embedding model name to use if none is provided.
-  static const defaultEmbeddingModelName = 'text-embedding-3-small';
+  }) : apiKey = apiKey ?? platform.getEnv(apiKeyName);
 
   /// The name of the environment variable that contains the API key.
   static const apiKeyName = 'OPENAI_API_KEY';
 
   @override
-  final String displayName = 'openai';
+  String get name => 'openai';
 
   @override
   final String? alias;
 
   /// The name of the OpenAI model to use.
-  final String modelName;
+  final String? modelName;
 
   /// The name of the OpenAI embedding model to use.
-  final String embeddingModelName;
+  final String? embeddingModelName;
 
   /// The API key to use for authentication with the OpenAI API.
   final String apiKey;
@@ -78,5 +70,5 @@ class OpenAiProvider extends Provider {
   );
 
   @override
-  final Iterable<ProviderCaps> caps;
+  final Set<ProviderCaps> caps;
 }