+++ lib/src/providers/interface/provider_caps.dart
@@ -10,17 +10,21 @@ enum ProviderCaps {
   chat,
 
   /// Indicates that the provider supports file uploads.
-  fileUploads;
+  fileUploads,
+
+  /// Indicates that the provider supports tool calls.
+  tools;
 
   /// All available capabilities
-  static const Iterable<ProviderCaps> all = [
+  static const Set<ProviderCaps> all = {
     textGeneration,
     embeddings,
     chat,
     fileUploads,
-  ];
+    tools,
+  };
 
   /// Returns all capabilities except those specified in [these].
-  static Iterable<ProviderCaps> allExcept(Iterable<ProviderCaps> these) =>
-      all.where((c) => !these.contains(c));
+  static Set<ProviderCaps> allExcept(Set<ProviderCaps> these) =>
+      all.where((c) => !these.contains(c)).toSet();
 }